ULAB Peer Mentorship/backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
pydantic==2.5.1
python-dotenv==1.0.0
motor==4.3.3
passlib[bcrypt]==1.7.5
pyjwt==2.9.0
